refactored the results
