+++ app/containers/App/index.js
@@ -9,8 +9,7 @@ import { ChevronLeft } from '@datapunt/asc-assets';
 import { useInjectSaga } from 'utils/injectSaga';
 
 import HomePage from 'containers/HomePage';
-import AddressInput from 'containers/AddressInput';
-import AddressInputResults from 'containers/AddressInput/Results';
+import Location from 'containers/Location';
 import QuestionnaireContainer from 'containers/QuestionnaireContainer';
 import AllQuestions from 'containers/QuestionnaireContainer/AllQuestions';
 import QuestionnaireRoutes from 'containers/QuestionnaireContainer/QuestionRoutes';
@@ -19,7 +18,7 @@ import Header from 'components/Header';
 import Footer from 'components/Footer';
 import GlobalError from 'containers/GlobalError';
 import questionnaireSaga from '../QuestionnaireContainer/saga';
-import addressInputSaga from '../AddressInput/saga';
+import addressInputSaga from '../Location/saga';
 
 import './style.scss';
 
@@ -87,11 +86,10 @@ export const App = () => {
               <Switch>
                 <Route exact path="/" component={HomePage} />
                 <Route exact path="/aanbouw/inleiding" component={HomePage} />
-                <Route exact path="/aanbouw/locatie" component={AddressInput} />
+                <Route exact path="/aanbouw/locatie" component={Location} />
                 <Route exact path="/aanbouw/alle-vragen" component={AllQuestions} />
                 <Route exact path="/aanbouw/alle-routes" component={QuestionnaireRoutes} />
                 <Route exact path="/aanbouw/*" component={QuestionnaireContainer} />
-                <Route exact path="/adres" component={AddressInputResults} />
                 <Route exact path="/health" />
                 <Route path="" component={NotFoundPage} />
               </Switch>
