+++ app/containers/Location/constants.js
