+++ app/containers/Location/reducer.js
