+++ app/containers/Location/reducer.test.js
