+++ app/containers/Location/selectors.js
